@@ -1,9 +1,13 @@
 {
-  "name": "anthill/searcher",
+  "name": "keltanas/searcher",
   "description": "parse query string",
   "minimum-stability": "stable",
   "license": "MIT",
   "authors": [
+    {
+      "name": "Nikolay Ermin",
+      "email": "keltanas@gmail.com"
+    },
     {
       "name": "Ivan Pyankov",
       "email": "unit.1985@gmail.com"
@@ -17,11 +21,21 @@
   },
 
   "require": {
-    "php": ">=5.3.23",
-    "symfony/event-dispatcher": "> 2.0",
+    "php": ">=7.0",
+    "symfony/event-dispatcher": "~2.8",
     "psr/log":"1.0.*"
   },
   "require-dev": {
-    "phpunit/phpunit": "~4.0"
+    "phpunit/phpunit": "~4.8"
+  },
+
+  "config": {
+    "bin-dir": "bin"
+  },
+
+  "extra": {
+    "branch-alias": {
+      "dev-master": "0.2-dev"
+    }
   }
 }
