@@ -4,8 +4,10 @@ namespace Searcher\LoopBack\Parser;
 
 
 use Searcher\LoopBack\Parser\Filter\Condition\Exception\InvalidConditionException;
+use Searcher\LoopBack\Parser\Filter\FilterGroupBuilder;
 use Searcher\LoopBack\Parser\Filter\FilterGroupConditionBuilder;
 use Searcher\LoopBack\Parser\Order\Order;
+use Searcher\LoopBack\Parser\Order\OrderBuilder;
 use Searcher\LoopBack\Parser\Pagination\LimitBuilder;
 use Searcher\LoopBack\Parser\Pagination\OffsetBuilder;
 use Searcher\StringUtils;
@@ -13,91 +15,73 @@ use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 
 class Builder implements BuilderInterface
 {
-
-    private static $mapping = array(
-        self::FILTER_WHERE => array(
-            "className" => "\\Searcher\\LoopBack\\Parser\\Filter\\FilterGroupBuilder",
-            "field" => "filters"
-        ),
-        self::FILTER_ORDER => array(
-            "className" => "\\Searcher\\LoopBack\\Parser\\Order\\OrderBuilder",
-            "field" => "orders"
-        ),
-        self::FILTER_LIMIT => array(
-            "className" => "\\Searcher\\LoopBack\\Parser\\Pagination\\LimitBuilder",
-            "field" => "limit"
-        ),
-        self::FILTER_OFFSET => array(
-            "className" => "\\Searcher\\LoopBack\\Parser\\Pagination\\OffsetBuilder",
-            "field" => "offset"
-        ),
-    );
-
     const FILTER_WHERE = 'where';
     const FILTER_ORDER = 'order';
     const FILTER_LIMIT = 'limit';
     const FILTER_OFFSET = 'skip';
 
-    /**
-     * @var
-     */
-    private $conditions;
-    /**
-     * @var EventDispatcherInterface
-     */
+    private static $mapping = [
+        self::FILTER_WHERE => [
+            'className' => FilterGroupBuilder::class,
+            'field' => 'filters'
+        ],
+        self::FILTER_ORDER => [
+            'className' => OrderBuilder::class,
+            'field' => 'orders'
+        ],
+        self::FILTER_LIMIT => [
+            'className' => LimitBuilder::class,
+            'field' => 'limit'
+        ],
+        self::FILTER_OFFSET => [
+            'className' => OffsetBuilder::class,
+            'field' => 'offset'
+        ],
+    ];
+
+    /** @var EventDispatcherInterface */
     private $dispatcher;
 
-    public function __construct($conditions, EventDispatcherInterface $dispatcher = null)
+    /** @var LimitBuilder */
+    private $limit = LimitBuilder::LIMIT_DEFAULT;
+
+    /** @var OffsetBuilder */
+    private $offset = OffsetBuilder::OFFSET_DEFAULT;
+
+    /** @var Order[] */
+    private $orders = [];
+
+    /** @var FilterGroupConditionBuilder[] */
+    private $filters = [];
+
+    public function __construct(EventDispatcherInterface $dispatcher = null)
     {
-        $this->conditions = $conditions;
         $this->dispatcher = $dispatcher;
     }
 
     /**
-     * @var LimitBuilder
+     * {@inheritdoc}
      */
-    private $limit = LimitBuilder::LIMIT_DEFAULT;
-
-    /**
-     * @var OffsetBuilder
-     */
-    private $offset = OffsetBuilder::OFFSET_DEFAULT;
-
-    /**
-     * @var Order[]
-     */
-    private $orders = array();
-
-    /**
-     * @var FilterGroupConditionBuilder[]
-     */
-    private $filters = array();
-
-    /**
-     * @return $this
-     * @throws InvalidConditionException
-     */
-    public function build()
+    public function build($conditions = null)
     {
-        if (!is_array($this->conditions)) {
+        if (!is_array($conditions)) {
             return $this;
         }
 
-        foreach ($this->conditions as $key => $values) {
+        foreach ($conditions as $key => $values) {
             $key = StringUtils::toLower($key);
             if (!isset(self::$mapping[$key])) {
                 continue;
             }
 
-            $className = self::$mapping[$key]["className"];
+            $className = self::$mapping[$key]['className'];
 
-            $data = forward_static_call_array(
-                array($className, "create"),
-                array($values, $this->dispatcher)
-            );
+            /** @var BuilderInterface $builder */
+            $builder = $className::{'create'}($values, $this->dispatcher);
 
-            $fieldName = self::$mapping[$key]["field"];
-            $this->$fieldName = $data;
+            $fieldName = self::$mapping[$key]['field'];
+
+            $this->$fieldName = $builder;
         }
 
         return $this;
