@@ -8,61 +8,42 @@ use Searcher\LoopBack\Parser\Filter\Condition\Exception\InvalidConditionExceptio
 use Searcher\StringUtils;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 
-class FilterGroupBuilder implements BuilderInterface
+class FilterGroupBuilder implements BuilderInterface, FilterInterface
 {
-    /**
-     * @var EventDispatcherInterface
-     */
+    /** @var EventDispatcherInterface */
     private $dispatcher = null;
 
-    /**
-     * @param EventDispatcherInterface $dispatcher
-     */
+    /** @var FilterGroupConditionBuilder[] */
+    private $groups = [];
+
+    /** @param EventDispatcherInterface $dispatcher */
     public function setDispatcher(EventDispatcherInterface $dispatcher = null)
     {
         $this->dispatcher = $dispatcher;
     }
 
-    /**
-     * @var array
-     */
-    private $conditions = array();
-
-    /**
-     * @var FilterGroupConditionBuilder[]
-     */
-    private $groups = array();
-
     private function getAvailableGroups()
     {
-        return array(
+        return [
             FilterCondition::CONDITION_AND,
             FilterCondition::CONDITION_OR
-        );
+        ];
     }
 
     /**
-     * @param  $conditions
-     */
-    public function setGroupCondition($conditions)
-    {
-        $this->conditions = $conditions;
-    }
-
-    /**
+     * @inheritdoc
      * @return FilterGroupConditionBuilder[]
      */
-    public function build()
+    public function build($conditions = null)
     {
-        $groupConditions = $this->conditions;
-        $groups = array();
+        $groups = [];
 
-        foreach ($groupConditions as $groupOperator => $filterData) {
+        foreach ($conditions as $groupOperator => $filterData) {
             $group = StringUtils::toLower($groupOperator);
 
             if (!in_array($group, $this->getAvailableGroups())) {
                 $group = FilterCondition::CONDITION_AND;
-                $filterData = array($groupOperator => $filterData);
+                $filterData = [$groupOperator => $filterData];
             }
 
             $conditionBuilder = new FilterGroupConditionBuilder();
@@ -92,8 +73,7 @@ class FilterGroupBuilder implements BuilderInterface
     public static function create($groups, EventDispatcherInterface $dispatcher = null)
     {
         $instance = new static();
-        $instance->setGroupCondition($groups);
         $instance->setDispatcher($dispatcher);
-        return $instance->build();
+        return $instance->build($groups);
     }
 }
