@@ -10,85 +10,83 @@ class TransformerSearchTest extends \PHPUnit_Framework_TestCase
 
     public function testElasticTransformer()
     {
-        $inputQuery = array(
-            'where' => array(
-                'nin' => array('user_id' => array(3)),
-                'neq' => array('product_id' => '27'),
-                'or' => array(
-                    'lte' => array(
-                        'details.id' => 27
-                    ),
-                    'gte' => array(
-                        'details.id' => 10
-                    ),
-                    'in' => array('user_id' => array(1)),
-                ),
-                'like' => array('description' => 'My awe')
-            ),
-            'order' => array(
-                'user_id' => 'asc'
-            )
-        );
+        $inputQuery = [
+            'where' => [
+                'user_id' => ['nin' => [3]],
+                'product_id' => ['neq' => '27'],
+                'or' => [
+                    'details.id' => [
+                        'lte' => 27,
+                        'gte' => 10,
+                    ],
+                    'user_id' => ['in' => [1]],
+                ],
+                'description' => ['like' => 'My awe'],
+            ],
+            'order' => [
+                'user_id' => 'asc',
+            ],
+        ];
 
-        $builder = new Builder($inputQuery);
+        $builder = new Builder();
         $esBuilder = new ElasticSearchTransformer($builder);
-        $result = $esBuilder->build()->transform();
-        $this->assertEquals(array('size' => 25, 'from' => 0), $result);
+        $result = $esBuilder->build()->getResult();
+        $this->assertEquals(['size' => 25, 'from' => 0], $result);
 
+        $builder->build($inputQuery);
 
-        $expect = array(
+        $expect = [
             'from' => 0,
             'size' => 25,
-            'body' => array(
-                'sort' => array(
-                    array(
-                        'user_id' => 'asc'
-                    )
-                ),
-                'query' => array(
-                    'filtered' => array(
-                        'filter' => array(
-                            'bool' => array(
-                                'must_not' => array(
-                                    array(
-                                        'terms' => array(
-                                            'user_id' => array(3),
-                                        ),
-                                    ),
-                                    array(
-                                        'term' => array('product_id' => 27),
-                                    ),
-                                ),
-                                'should' => array(
-                                    array(
-                                        'range' => array(
-                                            'details.id' => array('lte' => 27),
-                                        ),
-                                    ),
-                                    array(
-                                        'range' => array(
-                                            'details.id' => array('gte' => 10),
-                                        ),
-                                    ),
-                                ),
-                                'must' => array(
-                                    array(
-                                        'regexp' => array(
-                                            'description' => array(
+            'body' => [
+                'sort' => [
+                    [
+                        'user_id' => 'asc',
+                    ],
+                ],
+                'query' => [
+                    'filtered' => [
+                        'filter' => [
+                            'bool' => [
+                                'must_not' => [
+                                    [
+                                        'terms' => [
+                                            'user_id' => [3],
+                                        ],
+                                    ],
+                                    [
+                                        'term' => ['product_id' => 27],
+                                    ],
+                                ],
+                                'should' => [
+                                    [
+                                        'range' => [
+                                            'details.id' => ['lte' => 27],
+                                        ],
+                                    ],
+                                    [
+                                        'range' => [
+                                            'details.id' => ['gte' => 10],
+                                        ],
+                                    ],
+                                ],
+                                'must' => [
+                                    [
+                                        'regexp' => [
+                                            'description' => [
                                                 'value' => 'My awe.*',
-                                            )
-                                        )
-                                    ),
-                                )
-                            ),
-                        ),
-                    ),
-                ),
-            )
-        );
-        $esBuilder = new ElasticSearchTransformer($builder->build());
-        $result = $esBuilder->build()->transform();
+                                            ],
+                                        ],
+                                    ],
+                                ],
+                            ],
+                        ],
+                    ],
+                ],
+            ],
+        ];
+        $esBuilder = new ElasticSearchTransformer($builder);
+        $result = $esBuilder->build()->getResult();
         $this->assertEquals($expect, $result);
     }
-
 }
