@@ -3,7 +3,7 @@
 namespace Searcher\Transformer;
 
 
-interface TransformerInterface {
-
-    public function transform();
+interface TransformerInterface
+{
+    public function getResult();
 }
